22/02/18 End of Day Update
Most of the day was spent attempting to get the Product Page to work.
Currently, everything does EXCEPT the Photo Upload feature. The Product
Images will upload and be stored in the correct folder, but will
promptly not attach themselves to the page. They will not display at all
due to not being attached to a page.